+++ src/features/bookclub/hooks/useFetchBookClubList.ts
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { BookClub, BookClubParams } from '@/types/bookclubs';
 import { useQuery } from '@tanstack/react-query';
 import { bookClubs } from '@/api/book-club/react-query';
-import { DEFAULT_FILTERS } from '@/lib/constants/filters';
+import { DEFAULT_FILTERS } from '@/constants/filters';
 
 const useBookClubList = ({ initialData }: { initialData: BookClub[] }) => {
   const [filters, setFilters] = useState<BookClubParams>(DEFAULT_FILTERS);
